Agregue la duracion en la rutas por estaciones y ademas , estoy agregando eso para el aproximado de tiempo

+++ orders/dto/orderDetailDtoBase.cs
@@ -7,7 +7,7 @@ namespace fletesProyect.orders.dto
 {
     public class orderDetailDtoBase
     {
-        public int quantity { get; set; }
+        public long quantity { get; set; }
 
     }
 }
